pyproject.toml
[project]
name = "golf-game"
version = "0.1.0"
description = "PGA Tour betting analysis tool - SkyBet odds vs tournament performance"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "pandas>=2.1.0",
    "playwright>=1.40.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.25.0",
    "pyyaml>=6.0.0",
    "tabulate>=0.9.0",
    "scikit-learn>=1.3.0",
    "statsmodels>=0.14.0",
    "numpy>=1.24.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sse-starlette>=2.0.0",
    "fpdf2>=2.8.6",
]
